提交

+++ ipo-service-provider/src/main/java/com/yrdce/ipo/modules/sys/entity/IpoBallotNoInfo.java
@@ -3,76 +3,76 @@ package com.yrdce.ipo.modules.sys.entity;
 import java.util.Date;
 
 public class IpoBallotNoInfo {
-    private Short id;
+	private int id;
 
-    private String commodityid;
+	private String commodityid;
 
-    private Short ballotnostartlen;
+	private Short ballotnostartlen;
 
-    private Short ballotnoendlen;
+	private Short ballotnoendlen;
 
-    private String ballotno;
+	private String ballotno;
 
-    private Date createtime;
+	private Date createtime;
 
-    public IpoBallotNoInfo(Short id, String commodityid, Short ballotnostartlen, Short ballotnoendlen, String ballotno, Date createtime) {
-        this.id = id;
-        this.commodityid = commodityid;
-        this.ballotnostartlen = ballotnostartlen;
-        this.ballotnoendlen = ballotnoendlen;
-        this.ballotno = ballotno;
-        this.createtime = createtime;
-    }
+	public IpoBallotNoInfo(int id, String commodityid, Short ballotnostartlen, Short ballotnoendlen, String ballotno, Date createtime) {
+		this.id = id;
+		this.commodityid = commodityid;
+		this.ballotnostartlen = ballotnostartlen;
+		this.ballotnoendlen = ballotnoendlen;
+		this.ballotno = ballotno;
+		this.createtime = createtime;
+	}
 
-    public IpoBallotNoInfo() {
-        super();
-    }
+	public IpoBallotNoInfo() {
+		super();
+	}
 
-    public Short getId() {
-        return id;
-    }
+	public int getId() {
+		return id;
+	}
 
-    public void setId(Short id) {
-        this.id = id;
-    }
+	public void setId(int id) {
+		this.id = id;
+	}
 
-    public String getCommodityid() {
-        return commodityid;
-    }
+	public String getCommodityid() {
+		return commodityid;
+	}
 
-    public void setCommodityid(String commodityid) {
-        this.commodityid = commodityid == null ? null : commodityid.trim();
-    }
+	public void setCommodityid(String commodityid) {
+		this.commodityid = commodityid == null ? null : commodityid.trim();
+	}
 
-    public Short getBallotnostartlen() {
-        return ballotnostartlen;
-    }
+	public Short getBallotnostartlen() {
+		return ballotnostartlen;
+	}
 
-    public void setBallotnostartlen(Short ballotnostartlen) {
-        this.ballotnostartlen = ballotnostartlen;
-    }
+	public void setBallotnostartlen(Short ballotnostartlen) {
+		this.ballotnostartlen = ballotnostartlen;
+	}
 
-    public Short getBallotnoendlen() {
-        return ballotnoendlen;
-    }
+	public Short getBallotnoendlen() {
+		return ballotnoendlen;
+	}
 
-    public void setBallotnoendlen(Short ballotnoendlen) {
-        this.ballotnoendlen = ballotnoendlen;
-    }
+	public void setBallotnoendlen(Short ballotnoendlen) {
+		this.ballotnoendlen = ballotnoendlen;
+	}
 
-    public String getBallotno() {
-        return ballotno;
-    }
+	public String getBallotno() {
+		return ballotno;
+	}
 
-    public void setBallotno(String ballotno) {
-        this.ballotno = ballotno == null ? null : ballotno.trim();
-    }
+	public void setBallotno(String ballotno) {
+		this.ballotno = ballotno == null ? null : ballotno.trim();
+	}
 
-    public Date getCreatetime() {
-        return createtime;
-    }
+	public Date getCreatetime() {
+		return createtime;
+	}
 
-    public void setCreatetime(Date createtime) {
-        this.createtime = createtime;
-    }
+	public void setCreatetime(Date createtime) {
+		this.createtime = createtime;
+	}
 }
